+++ packages/pangea-node-sdk/src/config.ts
@@ -9,6 +9,7 @@ class PangeaConfig {
   requestTimeout: number = 5000;
   queuedRetryEnabled: boolean = true;
   queuedRetries: number = 4;
+  customUserAgent: string | undefined = "";
 
   constructor(options?: ConfigOptions) {
     Object.assign(this, options);