+++ packages/react-mui-audit-log-viewer/CHANGELOG.md
@@ -5,6 +5,30 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.0.33] - 2023-04-04
+
+### Fixed
+
+- Updated to display field id in table
+
+## [0.0.32] - 2023-03-28
+
+### Fixed
+
+- Updated to show max results
+
+## [0.0.30] - 2023-03-27
+
+### Fixed
+
+- Update @pangeacyber/react-mui-shared version to 0.0.13
+
+## [0.0.29] - 2023-03-22
+
+### Fixed
+
+- Fix columns not updating properly
+
 ## [0.0.28] - 2023-03-22
 
 ### Fixed